Refactors tekton pipeline reconciler to use tektoninstallerset client
similar to tekton triggers, removes duplicate code
