+++ pkg/reconciler/common/releases.go
@@ -18,7 +18,6 @@ package common
 
 import (
 	"fmt"
-	"io/ioutil"
 	"os"
 	"path"
 	"path/filepath"
@@ -112,7 +111,7 @@ func sanitizeSemver(version string) string {
 func allReleases(instance v1alpha1.TektonComponent) ([]string, error) {
 	// List all the directories available under kodata
 	pathname := ComponentDir(instance)
-	fileList, err := ioutil.ReadDir(pathname)
+	fileList, err := os.ReadDir(pathname)
 	if err != nil {
 		return nil, err
 	}
