+++ pkg/reconciler/kubernetes/tektonpipeline/controller.go
@@ -25,6 +25,7 @@ import (
 	tektonPipelineInformer "github.com/tektoncd/operator/pkg/client/injection/informers/operator/v1alpha1/tektonpipeline"
 	tektonPipelineReconciler "github.com/tektoncd/operator/pkg/client/injection/reconciler/operator/v1alpha1/tektonpipeline"
 	"github.com/tektoncd/operator/pkg/reconciler/common"
+	"github.com/tektoncd/operator/pkg/reconciler/kubernetes/tektoninstallerset/client"
 	"k8s.io/client-go/tools/cache"
 	kubeclient "knative.dev/pkg/client/injection/kube/client"
 	"knative.dev/pkg/configmap"
@@ -62,15 +63,15 @@ func NewExtendedController(generator common.ExtensionGenerator) injection.Contro
 		if err != nil {
 			logger.Fatal(err)
 		}
+		tisClient := operatorclient.Get(ctx).OperatorV1alpha1().TektonInstallerSets()
 
 		c := &Reconciler{
-			operatorClientSet: operatorclient.Get(ctx),
-			kubeClientSet:     kubeclient.Get(ctx),
-			extension:         generator(ctx),
-			manifest:          manifest,
-			operatorVersion:   operatorVer,
-			pipelineVersion:   pipelineVer,
-			metrics:           metrics,
+			kubeClientSet:   kubeclient.Get(ctx),
+			extension:       generator(ctx),
+			manifest:        manifest,
+			pipelineVersion: pipelineVer,
+			installerSetClient: client.NewInstallerSetClient(tisClient, &manifest,
+				operatorVer, pipelineVer, v1alpha1.KindTektonPipeline, filterAndTransform(generator(ctx)), metrics),
 		}
 		impl := tektonPipelineReconciler.NewImpl(ctx, c)
